Fix for when a form contains a input called "name"
When a form contains a input called name calling name on the form will return that input element, not the forms name. This results in a selector like `form[name=[object Object]]` when later stringified. The fix is to call getAttribute() instead of accessing the property directly.

Tested on Chrome 59.0.3071.109.